Carve out swc/soma_test.py and merge single-target vis and cell tests
Four moves, each filing tests under the module they actually call:

  io/swc/reader_test.py -> soma_test.py for the three tests that call
      is_contour_soma / is_special_three_point_soma directly, together with
      the _soma_row builder they exclusively use. The soma *geometry* tests
      assert SwcReader output rather than these predicates' return values,
      so they stay with the reader.

  _multi_compartment/cell_vis_discrete_test.py and cell_vis_node_test.py
      -> cell_test.py. Cell.vis_topology/vis_cv/vis_branch/vis_node all live
      in cell.py; there is no cell_vis.py to name a test file after. The two
      byte-identical _build_tree copies collapse into one _soma_dend_tree.

  vis/layout/_property_test.py -> layout/_dispatch_test.py. Despite the
      underscore name it is not a package guard: it drives
      build_layout_branches_2d, so _dispatch.py is its real target. The
      hypothesis availability probe and no-op `given` stub travel with it.

  vis/vis_pyvista_real_files_test.py -> backend_pyvista_test.py, kept as its
      own class since it is the only real-pyvista path in a file that is
      otherwise fake-plotter based.

FIXTURE_DIR / VALID_SWC_FIXTURES / ALLOWED_TYPES were duplicated verbatim
across the two vis real-file tests; they now live in vis/_testing.py.

Collected test count is unchanged at 2276.
