+++ braincell/io/swc/reader_test.py
@@ -24,7 +24,6 @@ import numpy as np
 
 from braincell import Morphology
 from braincell.io.swc import SwcReadOptions, SwcReader
-from braincell.io.swc.soma import is_contour_soma, is_special_three_point_soma
 from braincell.io.swc.types import _SwcAttach, _SwcBranch, _SwcRow
 
 FIXTURE_DIR = Path(__file__).resolve().parents[3] / "data" / "morphology"
@@ -52,27 +51,6 @@ class SwcReaderTest(unittest.TestCase):
     def _issue_codes(self, report) -> set[str]:
         return {issue.code for issue in report.issues}
 
-    def _soma_row(
-        self,
-        node_id: int,
-        x: float,
-        y: float,
-        z: float = 0.0,
-        radius: float = 5.0,
-        parent_id: int = -1,
-    ) -> _SwcRow:
-        return _SwcRow(
-            line_number=node_id,
-            fields=tuple(),
-            node_id=node_id,
-            type_code=1,
-            x=x,
-            y=y,
-            z=z,
-            radius=radius,
-            parent_id=parent_id,
-        )
-
     def _branch_points_um(self, branch) -> np.ndarray:
         if branch.points_proximal is None or branch.points_distal is None:
             raise AssertionError("Branch does not expose point geometry.")
@@ -84,59 +62,6 @@ class SwcReaderTest(unittest.TestCase):
         radii = [branch.radii_proximal[0].to_decimal(u.um)]
         radii.extend(radius.to_decimal(u.um) for radius in branch.radii_distal)
         return np.array(radii, dtype=float)
-
-    def test_contour_rule_is_disabled(self) -> None:
-        acute_rows = (
-            self._soma_row(1, 0.0, 0.0),
-            self._soma_row(2, 5.0, 1.0),
-            self._soma_row(3, 0.0, 2.0),
-            self._soma_row(4, -5.0, 1.0),
-        )
-        right_rows = (
-            self._soma_row(1, 0.0, 0.0),
-            self._soma_row(2, 5.0, 5.0),
-            self._soma_row(3, 8.0, 1.0),
-            self._soma_row(4, 10.0, 0.0),
-        )
-        obtuse_rows = (
-            self._soma_row(1, 0.0, 0.0),
-            self._soma_row(2, 3.0, 4.0),
-            self._soma_row(3, 7.0, 4.0),
-            self._soma_row(4, 10.0, 0.0),
-        )
-
-        self.assertFalse(is_contour_soma(acute_rows))
-        self.assertFalse(is_contour_soma(right_rows))
-        self.assertFalse(is_contour_soma(obtuse_rows))
-
-    def test_special_three_point_rule_uses_center_first_topology(self) -> None:
-        special_rows = (
-            self._soma_row(1, 0.0, 0.0, radius=5.0, parent_id=-1),
-            self._soma_row(2, -5.0, 0.0, radius=5.0, parent_id=1),
-            self._soma_row(3, 0.0, 5.0, radius=5.0, parent_id=1),
-        )
-        chain_rows = (
-            self._soma_row(1, -5.0, 0.0, radius=5.0, parent_id=-1),
-            self._soma_row(2, 0.0, 0.0, radius=5.0, parent_id=1),
-            self._soma_row(3, 5.0, 0.0, radius=5.0, parent_id=2),
-        )
-        empty_children = {1: [], 2: [], 3: []}
-
-        is_special, ordered = is_special_three_point_soma(special_rows, empty_children)
-        self.assertTrue(is_special)
-        self.assertIsNotNone(ordered)
-        self.assertEqual(ordered[0].node_id, 1)
-        self.assertFalse(is_special_three_point_soma(chain_rows, empty_children)[0])
-
-    def test_special_three_point_rule_requires_leaf_side_points(self) -> None:
-        rows = (
-            self._soma_row(1, 0.0, 0.0, radius=5.0, parent_id=-1),
-            self._soma_row(2, -5.0, 0.0, radius=5.0, parent_id=1),
-            self._soma_row(3, 0.0, 5.0, radius=5.0, parent_id=1),
-        )
-
-        self.assertFalse(is_special_three_point_soma(rows, {1: [2, 3], 2: [4], 3: []})[0])
-        self.assertFalse(is_special_three_point_soma(rows, {1: [2, 3], 2: [], 3: [4]})[0])
 
     def test_reader_builds_compressed_tree_and_maps_known_types(self) -> None:
         path = self._write_swc(
